allow customized menu and navigation
This feature enables to create a template file named
navmenu.html optionally.

If this exists, it will included into the menu div and
allows to extend the navigation for some custom
stativ pages.

Also if exist, there is no redirect from "/" to "/pkg/"
anymore. In case of this also a custom index page
could be created.

This patch enables the godoc server to a full customizable
documentation site of go projects.

PR #309 should be applied in addition so that only partial
overload of templates is necessary.